requirements.txt
# Core dependencies
fastapi>=0.116.1
uvicorn>=0.35.0
python-multipart>=0.0.20
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.43
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
# passlib 1.7.4 is not yet compatible with bcrypt >=4.0 (raises 72-byte errors)
bcrypt>=3.2.0,<4.0

# Document Processing
pytesseract>=0.3.13
tika>=3.1.0

pdf2image>=1.17.0
opencv-python>=4.12.0.88
Pillow>=10.0.0

# NLP & ML
transformers>=4.55.3
torch>=2.8.0
sentence-transformers>=5.1.2
scikit-learn>=1.7.2
pandas>=2.3.3
textblob>=0.18.0
datasets
trl>=0.26.1
peft>=0.18.0

# Redis
redis>=6.4.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0